output: Report space on the target filesystem
For output modes which use the common 'create_local_output_disks'
function (those that produce '*-sda' files), report how much space is
available on the filesystem hosting that output file.  This is because
we've had failures where there wasn't enough space on the target
filesystem, so this just collects the information ahead of time in
case we see those kinds of failures again.

The enhanced output looks like:

  libguestfs: trace: disk_create "/var/tmp/p/fedora-40-sda" "raw" 6442450944
  libguestfs: trace: disk_create = 0
  Filesystem             1K-blocks     Used Available Use% Mounted on
  /dev/mapper/tmp-vartmp 721361824 32674004 652998984   5% /var/tmp

Thanks: Martin Necas, Alex Kalenyuk