output: Add optional ?create parameter
This parameter of Output.create_local_output_disks and
Output.output_to_local_file allows the default creation of disks to be
skipped.  It defaults to true (ie. create them) which is the normal
behaviour.

This allows output modes to implement the '-oo create=false' option,
although none do so far.