v2v: Print the version of libnbd / nbdcopy in debug output
We didn't previously log the version of libnbd or nbdcopy, and nbdcopy
itself won't tell us (because we don't use the nbdcopy --verbose
option).  However this information can be useful to know.  This commit
enhances the verbose output of virt-v2v to include this.

With current versions of nbdcopy it will look like this:

  info: nbdcopy version:
  nbdcopy 1.23.4
  libnbd 1.23.4

Note the second line is the version of the installed library, which in
some (very unusual, and probably broken) situations might be different
from the version of the nbdcopy binary, which is why both are logged.

libnbd 1.23.5 will soon enhance the output to include the downstream
package version (such as the RPM version) which is even more useful.
In future the output will look like:

  info: nbdcopy version:
  nbdcopy 1.23.5 (libnbd-1.23.5-1.fc43)
  libnbd 1.23.5 (libnbd-1.23.5-1.fc43)

See: https://gitlab.com/nbdkit/libnbd/-/merge_requests/28