convert: Model target boot device
SeaBIOS recently changed how it works so it no longer initializes all
disks at boot.  To get around this, for some Linux BIOS guests, we
will have to assign a boot order to the disks, with <boot order='1'>
for the disk that contains the GRUB bootloader, and higher boot orders
assigned to the other disks.

As the first step, model the target boot device.

In the current commit this is always unset (set to 'None'), so this
does nothing.

(cherry picked from commit e512d84bc8e18734aac7659e839f32c960f6fb0a)