v2v: Add --memsize and --smp options
For a long time in virt-builder and virt-customize, we have had
options --memsize and --smp which let you override the default amount
of memory and number of vCPUs assigned to the libguestfs appliance.
Usually virt-builder / virt-customize choose suitable numbers, but
occasionally (for example if you need to --install a large package or
--run a heavyweight program) we cannot anticipate what is needed so
it's useful to allow these to be overridden by the user.

Virt-v2v by default choses 2560MB and min (8, NR_CPUS) for these.
However much the same as above applies, especially since we added
virt-customize features to virt-v2v.  It's useful to have these
options as additional control or to "just get it working", although
generally we should still try to make virt-v2v choose good defaults.

Also add the same options to virt-v2v-in-place and virt-v2v-inspector.