input, output: Set nbdkit --name matching the socket name
If nbdkit supports it, set the nbdkit --name parameter to match the
socket name (usually "inX" or "outX").

Debug messages look like:

  nbdkit[in0]: debug: nbdkit 1.45.6
  nbdkit[in0]: debug: TLS disabled: could not load TLS certificates
  nbdkit[in0]: debug: service mode: unix socket
  nbdkit[in0]: debug: NBD URI: nbd+unix://?socket=/tmp/v2v.UEjC18/in0

For the two cases where we use a shared nbdkit (using export names to
select the disk), I used "in" and "out" for the name, rather than the
socket name.  That's slightly clearer.

This leaves out -o ovirt-upload as it's complicated to modify the
code, and this is now a legacy case.