build: Allow -o glance output mode to be disabled at build time
Use ./configure --disable-glance to disable the -o glance output mode.
We will use this in RHEL builds since in RHEL we don't support this
method of importing guests to OpenStack.

  $ virt-v2v -i disk /var/tmp/fedora-42.img -o glance
  virt-v2v: error: failure: -o glance is not enabled in this build

'virt-v2v --machine-readable' does not list 'output:glance', and the
existing test is conditionalized on this.

The documentation for -o glance is also excluded, using some hairy m4
preprocessing.  If modifying the m4 conditionals, check very carefully
that the output is exactly what you expect, since m4 quoting is very
hit-and-miss.  This reference was helpful: https://mbreen.com/m4.html

This commit is based on the downstream commit that we include in
RHEL 10.1:

https://github.com/libguestfs/virt-v2v/commit/5c5d4163dfa7cbdd10450d677a85aeaf87e1a14b