-o libvirt: don't output floppy XML when qemu lacks support
RHEL10 qemu doesn't support floppy devices. Currently if virt-v2v
is converting to libvirt output, and the input has a floppy device,
the generated XML will be invalid on RHEL10 host.

Teach virt-v2v to check for floppy support in domain capabilities.
If support is missing we never output a floppy device, and instead
log a warning when the input requests floppy.

Fixes: https://issues.redhat.com/browse/RHEL-90175

Signed-off-by: Cole Robinson <crobinso@redhat.com>