convert: Look for GRUB signature first to identify boot device
In commit ca6ec6317e ("convert: Detect target boot device for Linux
guests") we tried to identify the boot order by looking for the device
containing the /boot filesystem.  However during our testing we found
a guest where GRUB was installed on /dev/sda but the /boot partition
was found on /dev/sdd.  This guest would not boot with the boot order
indicating that disk 4 (sdd) was the boot disk.

Change how we search for the boot device by first trying to find a
GRUB signature in the boot sector of any disk.  The first matching
disk (if any) would then be picked as the boot device.  If this fails
then we fall back to looking at the device containing /boot as before.

This also updates the common module to get:

  Richard W.M. Jones (1):
      mlstdutils: Export List.find_opt

Reported-by: Ming Xie
Thanks: Gerd Hoffmann
Fixes: commit ca6ec6317e20a633315f783a8ba4ece3c2fc01f2
Fixes: https://issues.redhat.com/browse/RHEL-108991
See-also: https://thestarman.pcministry.com/asm/mbr/GRUB.htm