docs: Add link to official docs for removing VMware Tools
VMware/Broadcom published an official article on how to remove VMware
Tools manually on non-VMware hypervisors.  The article seems to have
been published relatively recently.  The earliest reference I can find
is from the IA, in mid 2024.  The article contains a 2022 date, but I
can find no external links to it before 2024.

Link to that, above the link to the unofficial method we described
before.

Also clean up the documentation to point out VMware Tools is buggy,
it's not our problem.