lib/utils: Defer SELinux test until use
Previously virt-v2v always ran 'getenforce ... | grep' to test if
SELinux was enforcing every time it was run, even for things like
'virt-v2v --help'.  However the result is only used if we run nbdkit,
qemu-nbd or use the ovirt-upload method.  Running nbdkit is common,
the others less so.  Defer the command until the result is needed.