convert: windows: Add a comment about not using guestfs_case_sensitive_path
In theory we ought to use 'g#case_sensitive_path path', but it's not
possible (and the current code is correct) because the guest
filesystem is not mounted.  Add a comment about this so I understand
in future why it is like this.

Related: commit 205a8d7064415f85053fa4fa85d40a0496326526