input/input_vddk.ml: Handle subdirectories in nbdkit vddk export wildcard
When calling nbdkit vddk plugin, we have to construct an export
parameter, which is a wildcard that matches all of the filenames we
will request.  See:
https://libguestfs.org/nbdkit-vddk-plugin.1.html#PARAMETERS

The wildcard is checked using fnmatch(FNM_PATHNAME), so '/' characters
in the filename do *not* match '*' characters in the wildcard.  As a
result, we failed before when the VMware filename was a path with
subdirectories.

Attempt some hairy functional programming to calculate a suitable
wildcard for these cases.  This definitely won't handle the case where
we have filenames at different directory depths, because in fact
nbdkit itself cannot handle this case right now.

Reported-by: Ming Xie
Signed-off-by: Richard W.M. Jones <rjones@redhat.com>
Fixes: https://issues.redhat.com/browse/RHEL-121728
Related: commit 5461976e229873a203062848c0de30e70067b3fb
Fixes: commit 076727e55f4d4fed246097d3f89ebfe83e3de88f
Related: https://issues.redhat.com/browse/RHEL-102734