convert: windows: Fix ESP conversion if C:\Windows\Temp has alternate case
We found a Windows guest which had the expected C:\Windows\Temp path,
but the actual name was TEMP rather than Temp.  It's unclear if this
had been renamed, or if Windows can create this normally.  In any
case, the unexpected casing caused this failure:

  virt-v2v: error: libguestfs error: mkdtemp: /Windows/Temp/ESP_XXXXXX: No such file or directory

We have to use g#case_sensitive_path here since the directory is on
NTFS.

I also removed the hard-coded "\Windows" in favour of %systemroot%,
and fixed the indentation.

Fixes: https://issues.redhat.com/browse/RHEL-124569
Reported-by: Ulhas Surse