docs: Add more description for --memsize option
The adjustable --memsize option, which increases the amount of memory
available in the appliance, is necessary to work around at least one
unfixable setfiles / glibc bug.  Document in more detail what this
option does in a new section, and mention this issue.  (As the issue
can't be fixed properly, this is a documentation-only fix).

Fixes: https://issues.redhat.com/browse/RHEL-125116