Add Cinder AZ support for OpenStack output
Some OpenStack clouds require an availability zone to be set when creating a
volume. This change adds an availability-zone -oo parameter to the OpenStack output
to allow defining the availability zone for volume creation.

An issue with unquoted parameter values containing spaces in generated OpenStack
CLI calls was also found while testing this and resolved.