RHEL: output/output.ml: Remove reduce-memory-pressure=on (cache=none)
Virt-v2v tries to reduce the amount of page cache used when writing
out the disk image to the target storage. It does this by enabling the
reduce-memory-pressure=on option (previously cache=none) option in
nbdkit-file-plugin.

However, use of nbdkit-file-plugin cache=none option causes data
corruption with Dell PowerMax 8000 storage (only).

As this is only an advisory setting, remove it for now.

Fixes: https://issues.redhat.com/browse/RHEL-135617